Change 'tutorial' to 'tutor' and drop 'tutorials'
This is to make way for de-emphasizing the general tutorial capability
in the workbench.  The alias 'tutorial' now brings up getting-started
(the alias is defined in the start plugin, as is 'getting started'
itself).  We still want the tutorial plugin because it supports our
own tutorials.  https://github.com/nimbella-corp/workbench/issues/287